⚡ Reduce _LD_SliceIndex CPU calls and use integer
They only need to be set once per LOD slice (minor performance
improvement). And there were several cases where they were set as
floats.

+++ crest/Assets/Crest/Crest/Scripts/Underwater/UnderwaterRenderer.Effect.cs
@@ -374,7 +374,7 @@ namespace Crest
             // We sample shadows at the camera position. Pass a user defined slice offset for smoothing out detail.
             Helpers.SetShaderInt(underwaterPostProcessMaterial, ShaderIDs.s_CrestDataSliceOffset, dataSliceOffset, setGlobalShaderData);
             // We use this for caustics to get the displacement.
-            underwaterPostProcessMaterial.SetFloat(LodDataMgr.sp_LD_SliceIndex, 0);
+            underwaterPostProcessMaterial.SetInt(LodDataMgr.sp_LD_SliceIndex, 0);
 
             LodDataMgrAnimWaves.Bind(underwaterPostProcessMaterialWrapper);
             LodDataMgrSeaFloorDepth.Bind(underwaterPostProcessMaterialWrapper);
